// ecommerce/src/app/shared/interface/theme.interface.ts
export interface IHomeBanner {
  background_image?: string;
  status: boolean;
  banners: IBanners[];
}
// ...
export interface IProductTabSection {
  tag?: string;
  title?: string;
  sub_title?: string;
  description?: string;
  image_url?: string;
  category_ids: number[];
  status: boolean;
}

export interface IService {
  status: boolean;
  banners: IBanners[];
}

export interface ISocialMedia {
  title?: string;
  status: boolean;
  banners: IBanners[];
}

export interface IBrands {
  brand_ids: number[];
  status: boolean;
}
// ...
export interface ISliderProduct {
  status: boolean;
  tag?: string;
  title?: string;
  image_url?: string;
  product_slider_1: IProductSection;
  product_slider_2: IProductSection;
  product_slider_3: IProductSection;
  product_slider_4?: IProductSection;
}
// ...
export interface IProductCategoryFashionFour {
  title?: string;
  category_ids: number[];
  status: boolean;
}
// ...
export interface IBooks {
  id: number;
  content: IBooksContent;
  slug: string;
}

export interface IBooksContent {
  home_banner: IHomeBanner;
  categories_1: IProductCategoryFashionFour;
  category_product: IProductTabSection;
  categories_2: IProductCategoryFashionFour;
  slider_products: ISliderProduct;
  products_list: IProductTabSection;
  everyday_casual: IProductTabSection;
  brand: IBrands;
  products_ids: number[];
}